--- GAD/tempCodeRunnerFile.py
@@ -1,5 +1,0 @@
-
-    detections=net.forward()
-    faceBoxes=[]
-    for i in range(detections.shape[2]):
-        confidence=detections[0,0,i,2]
